Molmo2 tests: drop out-of-scope Qwen3.5-VL test + dedup shared helpers
- Remove src/test/nn/vision/qwen3_5_vl_test.py: it tests olmo_core.nn.vision.qwen3_5_vl,
  which is NOT on vision, so behind its skip-guards it only ever skips here. It belongs
  with the Qwen3.5-VL feature PR that adds the module (it rode along in the Molmo2 stage-1
  PR by mistake).
- Consolidate the duplicated _hf_cache_has (was defined identically in 3+ files) and
  MOLMO2_VARIANTS (2 files) into src/test/nn/vision/_molmo2_common.py; the parity/generation
  tests now import one copy. Pure dedup, no behavior change.

Net -194 lines.
