+++ src/test/nn/vision/molmo2_logits_parity_test.py
@@ -37,20 +37,7 @@ from olmo_core.testing import requires_gpu
 
 transformers = pytest.importorskip("transformers")
 
-MOLMO2_VARIANTS = [
-    "allenai/Molmo2-4B",
-    "allenai/Molmo2-8B",
-    "allenai/Molmo2-O-7B",
-]
-
-
-def _hf_cache_has(model_id: str) -> bool:
-    suffix = "models--" + model_id.replace("/", "--")
-    candidates = [os.path.expanduser("~/.cache/huggingface/hub")]
-    hf_home = os.environ.get("HF_HOME")
-    if hf_home:
-        candidates.append(os.path.join(hf_home, "hub"))
-    return any(os.path.isdir(os.path.join(root, suffix)) for root in candidates if root)
+from ._molmo2_common import MOLMO2_VARIANTS, _hf_cache_has  # noqa: F401 (re-exported)
 
 
 def _build_ours(model_id: str, device, dtype):
